backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv
google-generativeai
requests